+++ portfolio/src/main/java/com/google/sps/servlets/DeleteCommentServlet.java
@@ -1,0 +1,17 @@
+package com.google.sps.servlets;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/** Servlet that handles deleting comment content. */
+@WebServlet("/delete-comment")
+public class DeleteCommentServlet extends CommentServlet {
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+
+    }
+}
