+++ portfolio/src/main/java/com/google/sps/utils/Constants.java
@@ -1,7 +1,7 @@
 package com.google.sps.utils;
 
 public class Constants {
-    public final static String COMMENT_KEY = "Comment";
+    public final static String COMMENT_KIND = "Comment";
     public final static String COMMENT_COMMENTER = "commenter";
     public final static String COMMENT_CONTENT = "content";
     public final static String COMMENT_TIMESTAMP = "timestamp";
