--- portfolio/src/main/webapp/comment.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Leave Your Comment Here</title>
-</head>
-<body>
-
-</body>
-</html>
