+++ portfolio/src/main/webapp/script.js
@@ -81,50 +81,3 @@ function goBack() {
         history.back();
     }
 }
-
-/**
- * Fetches the response of "/data".
- */
-function loadAndShowComments() {
-    fetch("/list-comment").then(response => response.json()).then((json) => {
-        const div = document.getElementById("comments");
-        div.innerHTML = "";
-
-        for (let i = 0; i < json.length; i++) {
-            let commentString = getFormattedComment(json[i]);
-            div.appendChild(createListElement(commentString));
-        }
-    });
-}
-
-/** Creates an <li> element containing text. */
-function createListElement(text) {
-    const liElement = document.createElement('li');
-    liElement.innerText = text;
-    return liElement;
-}
-
-/**
- * Get a human readable string from a comment json.
- * @param {json} json Comment object in json form.
- * @return {string} A formatted string.
- */
-function getFormattedComment(json) {
-    let commenter = json.commenter;
-    let content = json.content;
-    let timestamp = json.timestamp;
-
-    let timeString = getFormattedDate(timestamp);
-
-    let resultString = `Commenter: ${commenter}\nTime: ${timeString}\nComment: ${content}`;
-
-    return resultString;
-}
-
-function getFormattedDate(timestamp) {
-    let date = new Date(timestamp);
-
-    let timeString = date.toLocaleString();
-
-    return timeString;
-}
